api/pyproject.toml
[project]
name = "opencorporates-deepwiki-api"
dynamic = ["version"]
description = "CLI tool and optional API server for OpenCorporates-DeepWiki, providing smart code analysis and AI-powered documentation generation."
license = {text = "MIT License"}

[tool.poetry]
name = "opencorporates-deepwiki-api"
version = "1.0.0"
description = "CLI tool and optional API server for OpenCorporates-DeepWiki, providing smart code analysis and AI-powered documentation generation."
authors = ["OpenCorporates DeepWiki Team"]
packages = [{include = "api", from = ".."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.95.0"
uvicorn = { extras = ["standard"], version = ">=0.21.1" }
pydantic = ">=2.0.0"
google-generativeai = ">=0.3.0"
tiktoken = ">=0.5.0"
adalflow = ">=0.1.0"
faiss-cpu = ">=1.7.4"
requests = ">=2.28.0"
python-dotenv = ">=1.0.0"
openai = ">=1.76.2"
ollama = ">=0.4.8"
aiohttp = ">=3.8.4"
boto3 = ">=1.34.0"
websockets = ">=11.0.3"
click = "^8.3.0"
enlighten = "^1.14.1"
simple-term-menu = "^1.6.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
deepwiki = "api.cli.main:cli"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
